+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to R Console will be documented in this file.
 
+## [0.3.2] - 2026-07-08
+
+### Added
+
+- Added vscode-R language-server routing for console completions through hidden `.rconsole` documents. This is to reduce extra R language-server processes and memory footprint when the shared route is selected.
+- Added `r.console.languageServer` to choose between vscode-R's language server and the console-owned language server.
+  - When `vscode-R` is selected, console completions use an exsiting language server started by vscode-R, or start one through vscode-R. This also respects the vscode-R's `r.lsp.multiServer`setting.
+  - When `console`is selected, the console starts its own separate language-server process.
+
 ## [0.3.1] - 2026-06-29
 
 ### Fixed